seprating content blacklist logic to controller and components
